refactor: 유저 패키지 분리
