+++ README.md
@@ -73,7 +73,7 @@
     - [x] 1~10, J, Q, K 중 하나를 랜덤으로 뽑는다.
     - [x] 프로그램 실행시 카드 모양과 카드 숫자를 조합하여 52장의 카드를 생성한다 
     - [x] 카드 배부 시 52장의 카드 덱에서 카드를 뽑아서 배부한다. (배부된 카드는 덱에서 지운다)
-- [ ] 모든 유저와 딜러는 랜덤으로 카드를 2장 받고 시작한다.
+- [x] 모든 유저와 딜러는 랜덤으로 카드를 2장 받고 시작한다.
 - [ ] 유저가 추가 배부를 시작했을 때
     - [ ] 추가로 받은 카드의 합이 21 초과 시 패배한다
     - [ ] 21 미만일 때 한번 더 배부받을 수 있다
