+++ src/main/java/domain/user/Player.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.user;
 
 public class Player extends User {
     public Player(String name) {
