+++ src/main/java/domain/user/User.java
@@ -1,4 +1,8 @@
-package domain;
+package domain.user;
+
+import domain.CardDeck;
+import domain.CardSetting;
+import domain.TrumpCard;
 
 public abstract class User {
     protected final String name;
